move files around, more refactoring
