+++ tests/workflow_test.go
@@ -22,7 +22,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package workflow
+package tests
 
 import (
 	"testing"
@@ -30,6 +30,6 @@ import (
 	"github.com/stretchr/testify/suite"
 )
 
-func TestWorkflowBufferedEventsTestSuite(t *testing.T) {
-	suite.Run(t, new(WorkflowBufferedEventsTestSuite))
+func TestWorkflowTestSuite(t *testing.T) {
+	suite.Run(t, new(WorkflowTestSuite))
 }
